Refactor services and reorganize project structure
Consolidated service implementations into dedicated files under `internal/service` for better modularity. Updated code generation paths and configurations, and replaced inline handlers with reusable service components. Adjusted project structure and workflows accordingly for consistency and maintainability.
