+++ Makefile
@@ -11,8 +11,8 @@ server-test:
 	@cd test && go test
 
 codegen:
-	 @rm -rf ./internal/test/gen/api.gen.go
-	 @oapi-codegen --config ./internal/test/resource/config.yaml ./internal/test/resource/pet-store.yaml
+	 @rm -rf ./internal/gen/api.gen.go
+	 @oapi-codegen --config ./internal/resource/config.yaml ./internal/resource/pet-store.yaml
 
 pet-store:
 	 @oapi-codegen --config ./test/resource/local-config.yaml ./test/resource/pet-store.yaml  2>&1 | pbcopy
