+++ internal/resource/pet-store.yaml
